+++ crates/test_file/src/lib.rs
@@ -1,0 +1,6 @@
+mod test_file;
+mod test_file_default;
+mod test_file_drop;
+mod test_file_from;
+
+pub use test_file::TestFile;
