+++ crates/test_file/src/test_file_default.rs
@@ -1,0 +1,8 @@
+use crate::TestFile;
+
+impl Default for TestFile {
+    #[track_caller]
+    fn default() -> Self {
+        Self::new()
+    }
+}
