+++ src/lib.rs
@@ -1,12 +1,9 @@
-extern crate db_error;
-
 mod db;
 mod graph;
 mod query;
 mod query_error;
 mod query_result;
 mod storage;
-mod test_utilities;
 mod transaction;
 
 pub use db::Db;
