--- src/test_utilities.rs
@@ -1,1 +1,0 @@
-pub(crate) mod test_file;
